Make it possible to compile the crate with wasm
There's no loading-from-function-pointers approach to creating the
GL(ow) context when targeting webGL, so this patch adds another API
variant to create the renderer from a give HTML Canvas element. That's
inspired by the glow API that also takes HTML elements sometimes and it
gives control over the context attributes to ensure that a stencil
buffer is present.