Fix text rendering with WebGL 1
The texture handling requires somes changes, due to the OpenGL ES 2.0
spec:

* In later versions gl::RED replaces gl::LUMINANCE, but with GLES 2.0
  we can only use the latter.
* The internal and external format must match with 2.0.
* The pixel store row parameters are not supported (maybe they can be
  removed?)

Finally the atlas texture is rendered into using an FBO. With 2.0 that
can't be backed by luminance texture, it needs to have all color
channels.

The previous "gles version" check was also replaced. The previous bool
was not really used and this patch changes it to be used to guard
exactly GLES 2.0 code paths. However detection through the version
string is not reliable, typically this is a compile time constant. So
instead we know the version when creating the WebGL 1 context and for
the regular API we assume it's not GLES 2.0, since previously 2.0 wasn't
supported anyway, so it cannot ever have worked.